MaestroController - Added default value checks when saving settings.
CueInterpreterDialog - Fixed Cue size.
DeviceControlWidget - Fixed Cue size.

+++ src/dialog/cueinterpreterdialog.cpp
@@ -5,7 +5,7 @@
 #include <QClipboard>
 
 namespace PixelMaestroStudio {
-	CueInterpreterDialog::CueInterpreterDialog(QWidget *parent, uint8_t* cuefile, uint16_t size) : QDialog(parent), ui(new Ui::CueInterpreterDialog), model_(cuefile, size) {
+	CueInterpreterDialog::CueInterpreterDialog(QWidget *parent, uint8_t* cuefile, uint32_t size) : QDialog(parent), ui(new Ui::CueInterpreterDialog), model_(cuefile, size) {
 		ui->setupUi(this);
 
 		ui->interpretedCueTableView->setModel(&model_);
